requirements.txt
pillow>=10.0.0
numpy